Рефактор SubjectDao и ISubjectDao

+++ src/main/java/ru/innopolis/stc9/susev/db/dao/SubjectDao.java
@@ -1,6 +1,7 @@
 package ru.innopolis.stc9.susev.db.dao;
 
 
+import com.sun.istack.internal.Nullable;
 import ru.innopolis.stc9.susev.db.connection.ConnectionManager;
 import ru.innopolis.stc9.susev.db.connection.ConnectionManagerJDBCImpl;
 import ru.innopolis.stc9.susev.pojo.Subject;
@@ -13,9 +14,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class SubjectDao implements ISubjectDao {
-    public static final String COLUMN_ID = "id";
-    public static final String COLUMN_SUBJECT_NAME = "subject_name";
-
     private ConnectionManager conManager = ConnectionManagerJDBCImpl.getInstance();
 
     @Override
@@ -23,56 +21,51 @@ public class SubjectDao implements ISubjectDao {
         if (subject == null) {
             return false;
         }
-        Connection connection;
-        connection = conManager.getConnection();
-        PreparedStatement statement = null;
-        statement = connection.prepareStatement("insert into subjects(id, subject_name) values (?, ?)");
-        statement.setInt(1, subject.getId());
-        statement.setString(2, subject.getName());
-
-        boolean execute = statement.execute();
-        connection.close();
-        return execute;
+        try (Connection connection = conManager.getConnection()) {
+            PreparedStatement statement = connection.prepareStatement(
+                    "insert into subjects(subject_name) values (?)");
+            statement.setString(1, subject.getName());
+            return statement.execute();
+        }
     }
 
     @Override
     public boolean deleteSubject(Subject subject) throws SQLException {
-        Connection connection;
-        connection = conManager.getConnection();
-        PreparedStatement statement = null;
-        statement = connection.prepareStatement("delete from subjects where id = ?");
-        statement.setInt(1, subject.getId());
-        boolean execute = statement.execute();
-        connection.close();
-        return execute;
+        if (subject == null) {
+            return false;
+        }
+        try (Connection connection = conManager.getConnection()) {
+            PreparedStatement statement = connection.prepareStatement("delete from subjects where id = ?");
+            statement.setInt(1, subject.getId());
+            return statement.execute();
+        }
     }
 
     @Override
+    @Nullable
     public Subject getSubjectById(int id) throws SQLException {
-        Subject subject = null;
-        Connection connection = ConnectionManagerJDBCImpl.getInstance().getConnection();
-        PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM subjects WHERE id=?");
-        preparedStatement.setInt(1, id);
-        ResultSet resultSet = preparedStatement.executeQuery();
-        if (resultSet.next()) {
-            subject = new Subject(resultSet.getInt(COLUMN_ID),
-                    resultSet.getString(COLUMN_SUBJECT_NAME));
+        try (Connection connection = conManager.getConnection()) {
+            PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM subjects WHERE id=?");
+            preparedStatement.setInt(1, id);
+            ResultSet resultSet = preparedStatement.executeQuery();
+            if (resultSet.next()) {
+                return getSubjectFromDb(resultSet);
+            }
+            return null;
         }
-        connection.close();
-        return subject;
     }
 
+    @Override
     public List<Subject> getSubjects() throws SQLException {
-        List<Subject> subjects = new ArrayList<>();
-        Connection connection = ConnectionManagerJDBCImpl.getInstance().getConnection();
-        PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM subjects");
-        ResultSet resultSet = preparedStatement.executeQuery();
-        while (resultSet.next()) {
-            subjects.add(new Subject(resultSet.getInt(COLUMN_ID),
-                    resultSet.getString(COLUMN_SUBJECT_NAME)));
+        try (Connection connection = conManager.getConnection()) {
+            List<Subject> subjects = new ArrayList<>();
+            PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM subjects");
+            ResultSet resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+                subjects.add(getSubjectFromDb(resultSet));
+            }
+            return subjects;
         }
-        connection.close();
-        return subjects;
     }
 
     @Override
@@ -80,17 +73,21 @@ public class SubjectDao implements ISubjectDao {
         if (subject == null) {
             return false;
         }
-        Connection connection;
-        connection = conManager.getConnection();
-        PreparedStatement statement = null;
-        statement = connection.prepareStatement("update subjects " +
-                "set subject_name = ?" +
-                "where id = ?");
-        statement.setString(1, subject.getName());
-        statement.setInt(2, subject.getId());
+        try (Connection connection = conManager.getConnection()) {
+            PreparedStatement statement = null;
+            statement = connection.prepareStatement("update subjects set subject_name = ? where id = ?");
+            statement.setString(1, subject.getName());
+            statement.setInt(2, subject.getId());
+            statement.executeUpdate();
+            return true;
+        }
+    }
 
-        statement.executeUpdate();
-        connection.close();
-        return true;
+    /**
+     * Get subject from DB
+     */
+    private Subject getSubjectFromDb(ResultSet set) throws SQLException {
+        return new Subject(set.getInt(COLUMN_ID),
+                set.getString(COLUMN_SUBJECT_NAME));
     }
 }
